requirements.txt
# ── Core ML ───────────────────────────────────
tensorflow>=2.13.0
numpy>=1.24.0
scikit-learn>=1.3.0

# ── Visualisation ──────────────────────────────
matplotlib>=3.7.0
seaborn>=0.12.0
Pillow>=10.0.0

# ── Streamlit UI ───────────────────────────────
streamlit>=1.28.0

# ── Utilities ──────────────────────────────────
tqdm>=4.66.0
